Don't assume that /boot/cmdline.txt exists

+++ setup/pi/make-root-fs-readonly.sh
@@ -19,14 +19,17 @@ function append_cmdline_txt_param() {
   # If the command line gets too long the pi won't boot.
   # Look for the option at the end ($) or in the middle
   # of the command line and surrounded by space (\s).
-  if ! grep -P -q "\s${toAppend}(\$|\s)" /boot/cmdline.txt
+  if [ -e /boot/cmdline.txt ] && ! grep -P -q "\s${toAppend}(\$|\s)" /boot/cmdline.txt
   then
     sed -i "s/\'/ ${toAppend}/g" /boot/cmdline.txt >/dev/null
   fi
 }
 
 function remove_cmdline_txt_param() {
-  sed -i "s/\(\s\)${1}\(\s\|$\)//" /boot/cmdline.txt > /dev/null
+  if [ -e /boot/cmdline.txt ]
+  then
+    sed -i "s/\(\s\)${1}\(\s\|$\)//" /boot/cmdline.txt > /dev/null
+  fi
 }
 
 log_progress "Disabling unnecessary service..."
